{"types":"C","runs":5}

+++ Readout.java
@@ -2,5 +2,5 @@
 public interface Readout
 {
 	public void update(String msg);
-	public void update(double data);
+	public void update(int data);
 }
